common/pyproject.toml
[project]
name = "common"
version = "0.1.0"
description = ""
authors = [{ name = "Trail of Bits", email = "opensource@trailofbits" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "protobuf (>=3.20, <=3.20.3)",
    "pydantic-settings>=2.7.1",
    "pymongo>=4.10.1",
    "nats-py>=2.8.0",
    "langchain-core~=0.3.32",
    "langchain-openai~=0.3.2",
    "langfuse ~= 2.59.2",
    "langchain ~= 0.3.18",
    "six>=1.17.0",
    "openlit == 1.32.12",                # 1.33 currently fails to import opentelemetry.sdk._events
    "langchain-community ~= 0.3.20",     # implicit dep of openlit if you instrument langchain
    "langchain-text-splitters ~= 0.3.7", # implicit dep of openlit if you instrument langchain
]

[project.scripts]
buttercup-util = "buttercup.common.util_cli:main"
buttercup-challenge-task = "buttercup.common.challenge_task_cli:main"
buttercup-task-registry = "buttercup.common.task_registry:task_registry_cli"


[tool.hatch.build.targets.wheel]
packages = ["src/buttercup"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.9.2",
    "mypy>=1.15.0",
    "types-requests",
    "types-PyYAML",
    "dirty-equals>=0.9.0",
]

[tool.ruff]
line-length = 120
exclude = [
    "./src/buttercup/common/datastructures",
    "./src/buttercup/common/clusterfuzz_env",
    "./src/buttercup/common/clusterfuzz_parser",
    "src/buttercup/common/clusterfuzz_utils.py",
]

[tool.mypy]
plugins = ["pydantic.mypy"]
mypy_path = "src"
packages = "buttercup"
allow_redefinition = true
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
sqlite_cache = true
strict_equality = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true
exclude = [
    "src/buttercup/common/clusterfuzz_parser",
    "src/buttercup/common/clusterfuzz_env",
    "src/buttercup/common/clusterfuzz_utils",
    "src/buttercup/common/datastructures",
]
